Python binding (#96)
* Python bindings for store

- Reorganizes interface files to mirror C++
- Adds store_old and raises NotImplementedError to specify non-C++
  method calls
- Impacted callers raise NotImplementedError to identify need for
  replacement to cpp version

* Operation and Constraints

* Purge python private details

* Partition

* Machine

* Cleanup

* Alternate cimport for github automation

* Remove old methods from docs

* Runtime

* Review

* Update docs with removed methods

* pre-commit

* Checkpoint for hello example

* Rename Cython store class

* Remove extra comments

* Compiler extras

* Store-to-array conversions to make tasks work

* Rename Store to LogicalStore so we can actually bind Store

* Binding for physical stores

* Fix the shutdown process

* Add missing files

* Minor fixes

* Inline allocations for Python

* Use public facing headers instead of detail

Validated against
- tests/unit
- examples/hello
- examples/transitivity

* Post-merge clean-up

* Stop using move constructors in Cython as they are not supported

* Overlap check for Logical stores

* Test a new naming convention

* Binding for scaling and bloating constraints

* Binding for indirect copies

* Uniform naming convention for Cython classes

* Stop heap-allocating C++ handles

* Workarounds for 0D stores

* Move LogicalArray to a separate Cython module

* Don't create a bogus tiling for 0D stores

* Support for the legacy communicator API

* Shutdown callback interface for Python

* Deprecate context APIs and rename Context to Library

* Reorganize Cython files to make them mirror the C++ counterparts

* Missing files

* Collapse Python and Cython modules for library and constraint

* Reorganize logical array, logical store, and types

* Recover the types module

* Make tasks stack allocable

* Move AutoTask completely to Cython

* Fix for examples

* Move ManualTask completely to Cython

* APIs to extract C++ handles from Cython objects

* Move Machine completely to Cython

* Minor fix

* Move Runtime to Cython

* Update examples to catch up the changes in APIs

* Missing file

* Fix the doc build

* Provenance tracking

* Binding for the tunable parameter API

* Fixes for typing stubs

* Move typing stubs for our own code back to the module directory

* Missing file for the sphinx doc

* Binding for point and rect types

* Export LEGATE_MAX_DIM

* Missing file

* Break cyclic imports

* Use stdint types in Cython modules

* Binding for the attaching variant of create_store

* Binding for broadcast constraints with no axes

* Use std::optional in the binding for legate::Slice

* Add fill by scalars to the binding and do some clean-up

* Fixes for overlap check and slicing

* Fix inline allocation for 0D stores

* More fixes for 0D stores

* Binding for detach

* Minor updates for the attach API

* Workarounds for post-shutdown destruction of Legion handles

* More workarounds

* Implement the array interface for inline allocations

* Cherry-pick the partitioner fix

* Clean up the binding code for Scalar

* Set the right logical region to a physical sub-store for domain calculation

* Non-nested provenance tracking

* Enable out-of-order destruction for logical stores in the Python binding

* Accept Scalar objects in add_scalar_arg

* Replace hand-written enum with a Cython generated one

* Remove unnecessary cmake calls

* Address review comments

* Fix the doc for machine APIs

* Address review comments

* Extract scalar store creation to a separate API

* Support construction of Scalars from bytes and ndarrays

* Another workaround for post-shutdown Legion calls

* Make sure attached regions get mapped in shared mode

* Remove dead code

* API to create stores from objects implementing the buffer protocol

* Fix compiler warnings

* Set the mapper argument correctly for image partitioning ops

* Typo

* Use initialized() to check the runtime shutdown in workarounds

* Ignore empty buffers in buffer builder

---------

+++ src/core/runtime/library.inl
@@ -65,9 +65,9 @@ void register_reduction_callback(const Legion::RegistrationCallbackArgs& args)
 namespace legate {
 
 template <typename REDOP>
-int32_t Library::register_reduction_operator(int32_t redop_id)
+int64_t Library::register_reduction_operator(int32_t redop_id)
 {
-  int32_t legion_redop_id = get_reduction_op_id(redop_id);
+  int64_t legion_redop_id = get_reduction_op_id(redop_id);
 #if !defined(REALM_COMPILER_IS_NVCC)
   if (LegateDefined(LEGATE_USE_CUDA)) {
     log_legate.error("Reduction operators must be registered in a .cu file when CUDA is enabled");
